Added ~20 additional SL rules.
Filled in details of the additional SL rules, leaving out only a handful that will require musical clarification to implement. Heavily "TODO" commented and will need to be updated once musical clarification happens. In addition, "TODO" commented in areas that may need to be altered if a structural revamp of this whole figured-bass-extractor package takes place (highly likely -- stay tuned!).