Added first batch of test files; fixed file input bug.
Added a folder of sample score blurbs to be used for testing the SL rules. These test musicxml files don't have a labeled 'bass' part, but the bass line is the lowest of 2 piano parts. To make the files accessible for testing, the lowest part in any score is now recognized as the bass part, if no labeled 'bass' part is present.

In addition, fixed a (one-word) bug that was only allowing scores already in the music21 corpus to be input, rather than any score the user has access to.

In addition, created an InputError exception in the work_browser class.